rest_api_backend_va_project/requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-debug-toolbar-force==0.1.8
django-environ==0.4.5
django-extensions==3.1.3
django-nine==0.2.4
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
Jinja2==3.0.1
jsonschema==3.2.0
loguru==0.5.3
MarkupSafe==2.0.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rfc3987==1.3.8
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.4
Werkzeug==2.0.1
win32-setctime==1.0.3
zope.interface==5.4.0
